my-first-excelsheet-agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my-first-excelsheet-agent"
version = "0.1.0"
description = "AI-Powered Google Sheets Chat Assistant"
authors = [
    { name = "Your Name" }
]
dependencies = [
    "streamlit>=1.30.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "google-generativeai>=0.3.0"
]

[tool.setuptools]
py-modules = ["app", "charts", "data_handler", "gemini_client", "ui_components"]